+++ package.json
@@ -42,6 +42,7 @@
   "license": "MIT",
   "dependencies": {
     "@material-ui/core": "^4.4.2",
+    "auth0-api-jwt-rsa-validation": "0.0.3",
     "axios": "^0.19.0",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.3.1",
